feat: Implement Lambertian Reflectance
This causes reflected rays to obey the distribution I = I0 cos(theta). This replaces the current reflectance distribution of uniform in all directions.

Visually this results in shadows being darker and more pronounced, since less light bounces towards the camera given that the directions of the normals in the shadow regions point away from the camera, and the light sections of the spheres being lighter and more blue given their increased scattering towards the sky.